refactor i18n loading: consolidate loadLocale function and remove i18nLoader.js
